+++ libraries/api-extractor/src/ast/AstStructuredType.ts
@@ -4,6 +4,8 @@
 /* tslint:disable:no-bitwise */
 
 import * as ts from 'typescript';
+import { ReleaseTag } from '../aedoc/ReleaseTag';
+import { ITextElement, IParagraphElement } from '../markup/OldMarkup';
 import AstMethod from './AstMethod';
 import AstProperty from './AstProperty';
 import AstItem, { AstItemKind, IAstItemOptions } from './AstItem';
@@ -155,6 +157,29 @@ export default class AstStructuredType extends AstItemContainer {
     return result;
   }
 
+  protected onCompleteInitialization(): void {
+    super.onCompleteInitialization();
+
+    // Is the constructor internal?
+    for (const member of this.getSortedMemberItems()) {
+      if (member.kind === AstItemKind.Constructor) {
+        if (member.documentation.releaseTag === ReleaseTag.Internal) {
+          // Add a boilerplate notice for classes with internal constructors
+          this.documentation.remarks.unshift(
+            {
+              kind: 'textDocElement',
+              value: `The constructor for this class is marked as internal. Third-party code`
+                + ` should not extend subclasses of the ${this.name} class or instantiate it directly.`
+            } as ITextElement,
+            {
+              kind: 'paragraphDocElement'
+            } as IParagraphElement
+          );
+        }
+      }
+    }
+  }
+
   private _processMember(memberSymbol: ts.Symbol, memberDeclaration: ts.Declaration): void {
     if (memberDeclaration.modifiers) {
       for (let i: number = 0; i < memberDeclaration.modifiers.length; i++ ) {
@@ -179,7 +204,7 @@ export default class AstStructuredType extends AstItemContainer {
     this._processedMemberNames.add(memberSymbol.name);
 
     const memberOptions: IAstItemOptions = {
-      extractor: this.extractor,
+      context: this.context,
       declaration: memberDeclaration,
       declarationSymbol: memberSymbol,
       jsdocNode: memberDeclaration