+++ core-build/gulp-core-build-mocha/CHANGELOG.md
@@ -1,6 +1,21 @@
 # Change Log - @microsoft/gulp-core-build-mocha
 
-This log was last generated on Thu, 05 Oct 2017 01:05:02 GMT and should not be manually modified.
+This log was last generated on Fri, 20 Oct 2017 19:57:12 GMT and should not be manually modified.
+
+## 3.1.5
+Fri, 20 Oct 2017 19:57:12 GMT
+
+*Changes not tracked*
+
+## 3.1.4
+Fri, 20 Oct 2017 01:52:54 GMT
+
+*Changes not tracked*
+
+## 3.1.3
+Fri, 20 Oct 2017 01:04:44 GMT
+
+*Changes not tracked*
 
 ## 3.1.2
 Thu, 05 Oct 2017 01:05:02 GMT